Move SingularTestNode and TestConfig
